refactor ars app code into ars_app module
