+++ tr_sys/tr_sys/urls.py
@@ -19,11 +19,12 @@ from django.urls import include, path
 from tr_ara_unsecret.unsecret_app import AppConfig as UnsecretApp
 from tr_ara_arax.arax_app import AppConfig as ARAXApp
 from tr_ara_bte.bte_app import AppConfig as BTEApp
+from tr_ars.default_ars_app.ars_app import AppConfig as ARSApp
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^ars/', include('tr_ars.urls')),
-    #url(r'^example/', include('tr_ars.default_ars_app.urls')),
+    #url(r'^example/', include(ARSApp.name)),
     url(r'^robokop/', include('tr_ara_robokop.urls')),
     url(BTEApp.regex_path, include(BTEApp.name)),
     url(r'^ncats/',include('tr_ara_ncats.urls')),
